Fix plugin for Rocket not working
